+++ client/src/messages.ts
@@ -4,7 +4,7 @@ const message = Cookies.get('message');
 if (message) {
     // Create a Bootstrap alert element
     const alertDiv = document.createElement('div');
-    alertDiv.className = 'alert alert-info position-fixed fade show text-nowrap fs-5 z-1';
+    alertDiv.className = 'alert alert-info position-fixed fade show text-nowrap fs-5 z-3';
     alertDiv.style.top = '50%';
     alertDiv.style.left = '50%';
     alertDiv.style.transform = 'translate(-50%, -50%)';