+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add privacy policy
+
 ### Changed
+
+- Make server log messages consistent
 
 ### Deprecated