+++ CHANGELOG.md
@@ -11,8 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Rooms are now only created via the `/create` paths; `/:room` will only join existing rooms
-
 ### Deprecated
 
 ### Removed
@@ -22,6 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 ---
+
+## [1.5.0] - 2025-09-10
+
+### Changed
+
+- Rooms are now only created via the `/create` paths; `/:room` will only join existing rooms
 
 ## [1.4.2] - 2025-09-09